refactor(core): 内聚类状态并消除伪解耦
